feat: self-injection 방식으로 내부 호출 문제 해결 방식 변경
<문제 상황>
findAllPosts()에서 getPostPage()를 내부 호출하기 때문에 getPostPage()의 @Cacheable 작동하지 않음

<기존 해결 방식>
별도 클래스에 getPostPage() 구현하여 findAllPosts()에서 호출

<기존 방식 문제점>
불필요하게 클래스 수 증가

<수정된 해결 방식>
PostService 에서 자기 자신을 주입하도록 하여 내부 호출시에도 @Cacheable 작동 되도록 함
